bump hc 145 hdk 139
