+++ Cargo.toml
@@ -2,7 +2,6 @@
 resolver = "2" # 2 is the default in 2021 edition
 members = [
   "zomes/hc_zome_trust_atom",
-  # "zomes/hc_zome_trust_graph",
 ]
 
 [profile.dev]
