+++ zomes/hc_zome_trust_atom/coordination/src/trust_atom.rs
@@ -28,11 +28,6 @@ const LINK_TAG_HEADER: [u8; 2] = [197, 166]; // Unicode "Ŧ" // hex bytes: [0xC5
 const LINK_TAG_ARROW_FORWARD: [u8; 3] = [226, 134, 146]; // Unicode "→" // hex bytes: [0xE2][0x86][0x92]
 const LINK_TAG_ARROW_REVERSE: [u8; 3] = [226, 134, 169]; // Unicode "↩" // hex bytes: [0xE2][0x86][0xA9]
 
-#[hdk_entry(id = "extra", visibility = "public")]
-#[derive(Clone)]
-pub struct Extra {
-  pub fields: BTreeMap<String, String>, // extra content
-}
 
 pub fn create(
   target: AnyLinkableHash,
