+++ zomes/hc_zome_trust_atom/integrity/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "hc_zome_tg_integrity"
+version = "0.0.10-dev"
+edition = "2021"
+rust-version = "1.56.0"
+
+[lib]
+name = "hc_zome_tg_integrity"
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+hdi = "0.0.14"
+rust_decimal = "1"
+serde = "1"
+derive_more = "0"
+
+[dev-dependencies]
+holochain = { version = "=0.0.145", default-features = false, features = [
+  "test_utils",
+] }
+tokio = { version = "1.3", features = ["full"] }
+futures = { version = "0.3.1", default-features = false }
